+++ backend/query/src/shared/data-with-count.ts
